requirements.txt
Django==4.2.30
djangorestframework==3.16.1
django-cors-headers==4.9.0
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.0.4
selenium==4.36.0
scikit-learn==1.6.1
numpy==2.0.2
PyMySQL==1.1.2
chromadb==1.5.7
